move extract_from_doc_chunks_function into domain
