+++ proscenium/scripts/document_enricher.py
@@ -14,8 +14,6 @@ from langchain_core.documents.base import Document
 from proscenium.verbs.chunk import documents_to_chunks_by_tokens
 from proscenium.verbs.extract import extract_to_pydantic_model
 
-# TODO move this method somewhere else
-
 
 def extract_from_document_chunks(
     doc: Document,
@@ -52,30 +50,6 @@ def extract_from_document_chunks(
     return extract_models
 
 
-def extract_from_doc_chunks_function(
-    doc_as_rich: Callable[[Document], Panel],
-    chunk_extraction_model_id: str,
-    chunk_extraction_template: str,
-    chunk_extract_clazz: type[BaseModel],
-    delay: float = 1.0,  # intra-chunk delay between inference calls
-    verbose: bool = False,
-) -> Callable[[Document], List[BaseModel]]:
-
-    def extract_from_doc_chunks(doc: Document):
-        chunk_extract_models = extract_from_document_chunks(
-            doc,
-            doc_as_rich,
-            chunk_extraction_model_id,
-            chunk_extraction_template,
-            chunk_extract_clazz,
-            delay,
-            verbose,
-        )
-        return chunk_extract_models
-
-    return extract_from_doc_chunks
-
-
 def enrich_documents(
     retrieve_documents: Callable[[], List[Document]],
     extract_from_doc_chunks: Callable[[Document], List[BaseModel]],
@@ -96,7 +70,7 @@ def enrich_documents(
 
             for doc in docs:
 
-                chunk_extract_models = extract_from_doc_chunks(doc)
+                chunk_extract_models = extract_from_doc_chunks(doc, verbose)
                 enrichments = doc_enrichments(doc, chunk_extract_models)
                 enrichments_json = enrichments.model_dump_json()
                 f.write(enrichments_json + "\n")
